restructure file system into a package
